Update README.md
This repository was created as part of a lab session to explore GitHub and Jupyter Notebooks. 

In this lab, we will:
- Set up and navigate GitHub
- Create and manage repositories
- Write and commit changes using Git
- Use Jupyter Notebooks to run Python code and document our work

The goal is to gain hands-on experience with version control and understand how GitHub supports collaboration in coding and data science projects. This repository will also serve as a workspace for experimenting with Python code in Jupyter.